SDK-178552: This change does not invoke L2 insert call if the L2 table is ful...
Devices: 56980_A0,56980_B0
Module: L2

Symptom:
SDK is raising DELETE callback notifications (BCM_L2_CALLBACK_DELETE)
whenever there is an internal hash movement due to hash multi move depth
configuration. And, it is seen that the ADD notifications are invoked
first and and then the DELETE notifications are invoked for the entries
that are internally moved. As a result, some of the MAC entries are
deleted in the application and so i could never scale more than ~7500
MAC entries

This change does not invoke L2 insert call if the L2 table is full. Only
when there is space in L2 table, we perform L2 table insertion. This
applies to both interrupt and polled modes. Any modification to L2 table
must be done through L2 APIs, or 'l2' command.