SDK-179560: To make a decision whether to make a delete related callback, per...
Devices: 56980_A0,56980_B0
Module: L2

Symptom:
L2 table is shuffled internally when making space for new entries,
handling collisions. When this process is taking place, the bcm sync
thread polls for changes in the L2 table. Since the entries can be moved
in to slots/locations which were previously unoccupied, the bcm sync
thread will interpret these as new entries and issue callbacks.

To make a decision whether to make a delete related callback, perform
table search. If entry is found (which will be at a different location),
do not issue callback. If entry is entry is not found, issue a (delete)
callback.